+++ src/components/music/sequence/style.tsx
@@ -15,6 +15,7 @@ export const StyledSequenceEditorWrapper = styled.div`
 
 interface StyledSequenceItemProps {
   $selected: boolean;
+  $filtered: boolean;
 }
 
 export const StyledSequenceItem = styled.div<StyledSequenceItemProps>`
@@ -36,6 +37,12 @@ export const StyledSequenceItem = styled.div<StyledSequenceItemProps>`
           outline: 4px solid ${(props) => props.theme.colors.highlight};
         `
       : ""}
+
+  ${(props) =>
+    props.$filtered &&
+    css`
+      filter: grayscale(1);
+    `}
 
   & .CustomSelect {
     max-width: 50px;